% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivcleanr.R
\docType{package}
\name{hivcleanr}
\alias{hivcleanr}
\alias{hivcleanr-package}
\title{hivcleanr: A package for cleaning HIV data.}
\description{
The hivcleanr package provides functions to clean and subset HIV data from OCA Excel-based tools.
}
\seealso{
\code{\link{oca_hiv_cleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
