% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oca_hiv_cleanr.R
\name{oca_hiv_cleanr}
\alias{oca_hiv_cleanr}
\title{OCA HIV data cleanr}
\usage{
oca_hiv_cleanr(x)
}
\arguments{
\item{x}{data frame containing raw programme data from OCA TB/HIV Excel tool (sheet = HIV_ART)}
}
\value{
data frame with variables renamed
}
\description{
Function to clean data from the OCA HIV Excel tool
}
\seealso{
\code{\link{hivcleanr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
